requirements.txt
debugpy==1.5.1
dill==0.3.6
einops==0.4.1
GPUtil==1.4.0
h5py==3.8.0
humanize==4.6.0
hydra-core==1.3.2
numpy==1.23.4
omegaconf==2.3.0
pandas==1.5.1
Pillow==9.5.0
psutil==5.9.0
python-dotenv==1.0.0
PyYAML==6.0
Requests==2.30.0
scikit_image==0.19.2
scikit_learn==1.2.1
timm==0.8.3.dev0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
wandb==0.15.3
torchmetrics==1.4.2
xformers==0.0.28
albumentations==1.4.17
pyarrow==17.0.0